+++ tests/c/syscalls/test_os_global_pin_is_validated.c
@@ -22,9 +22,39 @@
 hw_model_t hw_model = MODEL_COUNT;
 sdk_version_t sdk_version = SDK_COUNT;
 
+void *get_memory_code_address(void)
+{
+  return NULL;
+}
+
 struct app_s *get_current_app(void)
 {
   return NULL;
+}
+
+char *get_app_nvram_file_name(void)
+{
+  return NULL;
+}
+
+bool get_app_save_nvram(void)
+{
+  return false;
+}
+
+unsigned long get_app_nvram_address(void)
+{
+  return 0;
+}
+
+unsigned long get_app_nvram_size(void)
+{
+  return 0;
+}
+
+unsigned long get_app_text_load_addr(void)
+{
+  return 0;
 }
 
 void unload_running_app(bool UNUSED(unload_data))