+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] 2025-03-13
+
+### Added
+
+- `--save-nvram` and `--load-nvram` parameters that allow NVRAM testing
+- destination boundaries verification for `nvm_write()`
+
 ## [0.16.0] 2025-03-05
 
 ### Changed